+++ packages/components/src/stepper/Stepper.stories.tsx
@@ -1,4 +1,3 @@
-import { Input } from '@spark-ui/components/input'
 import { Tag } from '@spark-ui/components/tag'
 import { ArrowHorizontalDown } from '@spark-ui/icons/ArrowHorizontalDown'
 import { ArrowHorizontalUp } from '@spark-ui/icons/ArrowHorizontalUp'
@@ -28,9 +27,9 @@ const states: StepperProps['state'][] = ['error', 'alert', 'success']
 
 export const Default: StoryFn = _args => {
   return (
-    <Stepper>
+    <Stepper aria-label="Default stepper">
       <Stepper.DecrementButton aria-label="Decrement" />
-      <Stepper.Input aria-label="Default stepper" />
+      <Stepper.Input />
       <Stepper.IncrementButton aria-label="Increment" />
     </Stepper>
   )
@@ -38,9 +37,9 @@ export const Default: StoryFn = _args => {
 
 export const Uncontrolled: StoryFn = () => {
   return (
-    <Stepper defaultValue={0} onValueChange={console.log}>
+    <Stepper defaultValue={0} onValueChange={console.log} aria-label="Default stepper">
       <Stepper.DecrementButton aria-label="Decrement" />
-      <Stepper.Input aria-label="Uncontrolled stepper" />
+      <Stepper.Input />
       <Stepper.IncrementButton aria-label="Increment" />
     </Stepper>
   )
@@ -50,39 +49,39 @@ export const Controlled: StoryFn = () => {
   const [count, setCount] = useState(0)
 
   return (
-    <Stepper value={count} onValueChange={value => setCount(value ?? 0)}>
+    <Stepper value={count} onValueChange={setCount} aria-label="Default stepper">
       <Stepper.DecrementButton aria-label="Decrement" />
-      <Stepper.Input aria-label="Controlled stepper" />
+      <Stepper.Input />
       <Stepper.IncrementButton aria-label="Increment" />
     </Stepper>
   )
 }
 
 export const Disabled: StoryFn = _args => (
-  <Stepper disabled onValueChange={() => console.log('disabled')} defaultValue={10}>
+  <Stepper aria-label="Disabled stepper" disabled onValueChange={() => console.log('disabled')}>
     <Stepper.DecrementButton aria-label="Decrement" />
-    <Stepper.Input aria-label="Disabled stepper" />
+    <Stepper.Input />
     <Stepper.IncrementButton aria-label="Increment" />
   </Stepper>
 )
 
 export const ReadOnly: StoryFn = _args => (
-  <Stepper readOnly defaultValue={10}>
+  <Stepper aria-label="Readonly stepper" readOnly defaultValue={10}>
     <Stepper.DecrementButton aria-label="Decrement" />
-    <Stepper.Input aria-label="Readonly stepper" />
+    <Stepper.Input />
     <Stepper.IncrementButton aria-label="Increment" />
   </Stepper>
 )
 
 export const Custom: StoryFn = _args => (
-  <Stepper>
-    <Stepper.DecrementButton aria-label="Decrement">
+  <Stepper aria-label="Composed stepper">
+    <Stepper.DecrementButton aria-label="Decrement" design="filled" intent="basic">
       <Icon>
         <ArrowHorizontalDown />
       </Icon>
     </Stepper.DecrementButton>
-    <Stepper.Input aria-label="Composed stepper" />
-    <Stepper.IncrementButton aria-label="Increment">
+    <Stepper.Input />
+    <Stepper.IncrementButton aria-label="Increment" design="filled" intent="basic">
       <Icon>
         <ArrowHorizontalUp />
       </Icon>
@@ -90,41 +89,14 @@ export const Custom: StoryFn = _args => (
   </Stepper>
 )
 
-export const Standalone: StoryFn = _args => (
-  <div className="gap-lg flex flex-col items-start">
-    <div>
-      <Tag className="mb-md flex">No buttons</Tag>
-      <Stepper>
-        <Stepper.Input aria-label="Stepper without buttons" />
-      </Stepper>
-      <Input aria-label="Stepper without buttons" />
-    </div>
-
-    <div>
-      <Tag className="mb-md flex">Increment button only</Tag>
-      <Stepper>
-        <Stepper.Input aria-label="Stepper with only increment button" />
-        <Stepper.IncrementButton aria-label="Increment" />
-      </Stepper>
-    </div>
-    <div className="gap-md flex flex-col">
-      <Tag className="mb-md flex">Decrement button only</Tag>
-      <Stepper>
-        <Stepper.DecrementButton aria-label="Decrement" />
-        <Stepper.Input aria-label="Stepper with only decrement button" />
-      </Stepper>
-    </div>
-  </div>
-)
-
 export const State: StoryFn = _args => (
   <div className="gap-xl grid grid-cols-2 md:grid-cols-3">
     {states.map(state => (
       <div key={state}>
         <Tag className="mb-md flex">{state}</Tag>
-        <Stepper state={state}>
+        <Stepper aria-label={`Stepper ${state}`} state={state}>
           <Stepper.DecrementButton aria-label="Decrement" />
-          <Stepper.Input aria-label="Stepper with state" />
+          <Stepper.Input />
           <Stepper.IncrementButton aria-label="Increment" />
         </Stepper>
       </div>
@@ -133,17 +105,17 @@ export const State: StoryFn = _args => (
 )
 
 export const Step: StoryFn = _args => (
-  <Stepper step={10}>
+  <Stepper aria-label="Stepper with custom step" step={3}>
     <Stepper.DecrementButton aria-label="Decrement" />
-    <Stepper.Input aria-label="Stepper with custom step" />
+    <Stepper.Input />
     <Stepper.IncrementButton aria-label="Increment" />
   </Stepper>
 )
 
 export const MinMaxValues: StoryFn = _args => (
-  <Stepper minValue={0} maxValue={10} defaultValue={0}>
+  <Stepper aria-label="Stepper with min/max values" minValue={0} maxValue={100} defaultValue={0}>
     <Stepper.DecrementButton aria-label="Decrement" />
-    <Stepper.Input aria-label="Stepper with min/max values" />
+    <Stepper.Input />
     <Stepper.IncrementButton aria-label="Increment" />
   </Stepper>
 )
@@ -152,9 +124,14 @@ export const FormatOptions: StoryFn = _args => (
   <div className="gap-xl grid grid-cols-2 md:grid-cols-3">
     <div>
       <Tag className="mb-md flex">Percentages</Tag>
-      <Stepper minValue={0} defaultValue={0.01} formatOptions={{ style: 'percent' }}>
+      <Stepper
+        aria-label="Stepper with percentages"
+        minValue={0}
+        defaultValue={0.01}
+        formatOptions={{ style: 'percent' }}
+      >
         <Stepper.DecrementButton aria-label="Decrement" />
-        <Stepper.Input aria-label="Stepper with percentages" />
+        <Stepper.Input />
         <Stepper.IncrementButton aria-label="Increment" />
       </Stepper>
     </div>
@@ -162,21 +139,26 @@ export const FormatOptions: StoryFn = _args => (
     <div>
       <Tag className="mb-md flex">Currency values</Tag>
       <Stepper
+        aria-label="Stepper with currency value"
         defaultValue={1}
         step={5}
         formatOptions={{ style: 'currency', currency: 'EUR', currencyDisplay: 'symbol' }}
       >
         <Stepper.DecrementButton aria-label="Decrement" />
-        <Stepper.Input aria-label="Stepper with currency value" />
+        <Stepper.Input />
         <Stepper.IncrementButton aria-label="Increment" />
       </Stepper>
     </div>
 
     <div>
       <Tag className="mb-md flex">Units</Tag>
-      <Stepper defaultValue={20} formatOptions={{ style: 'unit', unit: 'celsius' }}>
+      <Stepper
+        aria-label="Stepper with units"
+        defaultValue={20}
+        formatOptions={{ style: 'unit', unit: 'celsius' }}
+      >
         <Stepper.DecrementButton aria-label="Decrement" />
-        <Stepper.Input aria-label="Stepper with units" />
+        <Stepper.Input />
         <Stepper.IncrementButton aria-label="Increment" />
       </Stepper>
     </div>
@@ -186,9 +168,9 @@ export const FormatOptions: StoryFn = _args => (
 export const WithFormField: StoryFn = _args => (
   <FormField name="title" isRequired state="error">
     <FormField.Label>Title</FormField.Label>
-    <Stepper minValue={0} maxValue={100} defaultValue={0}>
+    <Stepper aria-label="Stepper with min/max values" minValue={0} maxValue={100} defaultValue={0}>
       <Stepper.DecrementButton aria-label="Decrement" />
-      <Stepper.Input aria-label="Stepper with form field" />
+      <Stepper.Input />
       <Stepper.IncrementButton aria-label="Increment" />
     </Stepper>
     <FormField.ErrorMessage>oops</FormField.ErrorMessage>