+++ packages/components/src/stepper/index.ts
@@ -1,6 +1,9 @@
-import { Decrement as DecrementButton, Increment as IncrementButton } from './Buttons'
-import { Input } from './Input'
-import { Root } from './Root'
+import { Stepper as Root } from './Stepper'
+import {
+  StepperDecrementButton as DecrementButton,
+  StepperIncrementButton as IncrementButton,
+} from './StepperButton'
+import { StepperInput as Input } from './StepperInput'
 
 export const Stepper: typeof Root & {
   IncrementButton: typeof IncrementButton